requirements.txt
aiogram>=3.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
